# perf: optimize search module string operations and memory allocations