@medcl 2025-08-20
the old query logic will be removed soon, and also it feels like an over-optimization, and the unit tests are broken. I’d prefer not to proceed with it further.